Instantiate handler classes once
This commit changes the worker so that handler classes are only instantiated once. Before, each handler was instantiated every time it was invoked. This was OK for individual command handlers, but persistent handlers like TransactionLog or Proxy had to use the SharedState class to maintain state throughout the session.

With this change these classes are not persistent throughout the session. That means they can use class state (private and protected properties) to maintain their state. The change also means that we can load the handler configuration in __init__() rather than handle(). This makes the handler lifecycle much cleaner, methinks.