Add external content filter handler
This data handler allows you to run an external content filter to the data before delivery. The config file allows you to specify the command you want to run, the reject and flag numerical thresholds, and flags that you'll want to add to the message at the bottom of the header section. The external content filter command must return a numeric value. The handler will then compare the returned value to the thresholds and either reject the message, flag it, or leave it as-is.